fix app.__init__k.py

+++ app/__init__.py
@@ -4,7 +4,9 @@ logger_file_path = os.path.join(
     os.path.dirname(__name__),
     'example.log'
 )
-os.remove(logger_file_path)
+if os.path.exists(logger_file_path):
+    os.remove(logger_file_path)
+
 Logger = logging.basicConfig(
     filename=logger_file_path,
     level=logging.DEBUG
